backend/ai-service/requirements.txt
fastapi
uvicorn
pydantic
google-genai
google-auth
pandas
scikit-learn
joblib
numpy
xgboost
